# openapi.yaml
openapi: 3.1.0
info:
  title: IREC — API v1
  version: 1.0.0-draft.1
  description: |
    **BOZZA** — contratto di integrazione Mind → IREC, da concordare con il team Mind.
    Questo file è la fonte di verità dell'integrazione (contract-first).

    Convenzioni:
    - Autenticazione: call-token coniato da Mind (JWT firmato, verifica JWKS lato IREC).
      Claim richiesti: `sub`, `tenant_id`, `entitlement`, `scope`, `aud: "irec"`, `jti`, `exp`.
    - `x-correlation-id`: ricevuto da Mind e ri-emesso nei log di IREC.
    - Mutazioni idempotenti via header `Idempotency-Key`.
    - Run lunghe: risposta `202 { run_id }`, poi poll su GET.
    - Versioning `/v1`: solo cambi additivi; policy di deprecation da definire.
    - Gli endpoint a supporto dei tool dell'agente (addendum agentico §6:
      posizioni, storico solleciti, pausa/riprendi, ecc.) verranno aggiunti
      in modo additivo in revisioni successive di questa bozza.
servers:
  - url: /v1
security:
  - callToken: []
paths:
...
  /invoices/{invoice_id}/pause:
    post:
      operationId: pauseInvoice
      summary: Sospende il flusso su una fattura (scope irec.write)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: invoice_id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fino_a: { type: [string, 'null'], format: date }
                motivo: { type: string }
      responses:
        '200':
          description: Nuovo stato della fattura
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvoiceState' }
        '403': { $ref: '#/components/responses/EntitlementError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /invoices/{invoice_id}/resume:
    post:
      operationId: resumeInvoice
      summary: Riprende il flusso su una fattura in pausa (scope irec.write)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: invoice_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Nuovo stato della fattura
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvoiceState' }
        '403': { $ref: '#/components/responses/EntitlementError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /communications/{communication_id}/cancel:
    post:
      operationId: cancelCommunication
      summary: Annulla un singolo step programmato (scope irec.write)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: communication_id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/EntitlementError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /communications/{communication_id}/force:
    post:
      operationId: forceCommunication
      summary: Forza l'invio immediato di uno step (scope irec.write)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: communication_id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/EntitlementError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /invoices/{invoice_id}/payments:
    post:
      operationId: registerPayment
      summary: Registra un pagamento manuale, idempotente (scope irec.write)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: invoice_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [importo, data_pagamento, idempotency_key]
              properties:
                importo: { type: string, format: decimal }
                data_pagamento: { type: string, format: date }
                idempotency_key: { type: string }
      responses:
        '200':
          description: Esito del pagamento
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagamentoManuale' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/EntitlementError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /clients/{client_id}/contacts:
    patch:
      operationId: updateContacts
      summary: Aggiorna i recapiti di un cliente (scope irec.write)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: client_id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: [string, 'null'] }
                pec: { type: [string, 'null'] }
                telefono: { type: [string, 'null'] }
                canali_opt_out: { type: [array, 'null'], items: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/EntitlementError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /flow:
    put:
      operationId: replaceFlow
      summary: Sostituisce il flusso del mandante (Value/Premium, scope irec.write)
      description: >-
        La personalizzazione del flusso è disponibile dal pacchetto Value:
        un mandante Entry riceve 403 `upgrade_required` con invito garbato.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [steps]
              properties:
                steps:
                  type: array
                  items:
                    type: object
                    required: [ordine, offset_giorni, canale, template]
                    properties:
                      ordine: { type: integer }
                      offset_giorni: { type: integer }
                      canale: { type: string, enum: [email, pec, whatsapp, voice] }
                      template: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/EntitlementError' }
  /report:
    post:
      operationId: sendReport
      summary: Genera e invia il report al mandante via email (scope irec.write)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Esito dell'invio
          content:
            application/json:
              schema:
                type: object
                required: [inviato, destinatario_presente]
                properties:
                  inviato: { type: boolean }
                  destinatario_presente: { type: boolean }
        '403': { $ref: '#/components/responses/EntitlementError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }

  # Gli endpoint operativi vivono alla root, fuori dal prefisso /v1
  # (server override per-path).
...
components:
  securitySchemes:
    callToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Call-token coniato da Mind, verificato da IREC via JWKS.
  parameters:
    CorrelationId:
      name: x-correlation-id
      in: header
      schema: { type: string }
...
  responses:
    BadRequest:
      description: >-
        Richiesta non valida. `code`: validation_error,
        missing_idempotency_key, invalid_idempotency_key (oltre 128 caratteri).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ServiceUnavailable:
      description: >-
        Dipendenza non disponibile. `code`: auth_not_configured,
        jwks_unavailable, database_not_configured, database_unreachable,
        providers_not_configured.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    EntitlementError:
      description: >-
        Autorizzazione insufficiente. `code`: entitlement_missing (nessun
        entitlement IREC), scope_missing (scope richiesto assente dal token).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Risorsa non trovata
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: >-
        Stato incompatibile con l'azione. `code`: invoice_not_pausable,
        invoice_not_paused, invoice_not_active, invoice_already_paid,
        communication_not_cancellable, communication_not_forceable,
        channel_unavailable, run_in_progress.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Ok:
      description: Operazione eseguita
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
...
    InvoiceState:
      type: object
      required: [id, stato]
      properties:
        id: { type: string }
        stato: { type: string }
    PagamentoManuale:
      type: object
      required: [fattura_id, stato, gia_registrato, comunicazioni_annullate]
      properties:
        fattura_id: { type: string }
        stato: { type: string }
        gia_registrato: { type: boolean }
        comunicazioni_annullate: { type: integer }
...